src/show/program_options.cc: parse --position and --rotation as doubles
git-svn-id: svn+ssh://svn.code.sf.net/p/slam6d/code/trunk@1479 51f1e6e7-7289-49ad-8748-f09e43fc6daa

+++ src/show/program_options.cc
@@ -360,8 +360,8 @@ void validate(boost::any& v, const std::vector<std::string>& values,
   validators::check_first_occurrence(v);
   const char *s = validators::get_single_string(values).c_str();
 
-  int x, y, z;
-  if (sscanf(s, "%d,%d,%d", &x, &y, &z) == 3) {
+  double x, y, z;
+  if (sscanf(s, "%lf,%lf,%lf", &x, &y, &z) == 3) {
     v = boost::any(Position(x, y, z));
   } else {
     throw validation_error(validation_error::invalid_option_value, "Camera coordinates must be given as \"X,Y,Z\".");
@@ -376,8 +376,8 @@ void validate(boost::any& v, const std::vector<std::string>& values,
   validators::check_first_occurrence(v);
   const char *s = validators::get_single_string(values).c_str();
 
-  int x, y, z, w;
-  if (sscanf(s, "%d,%d,%d,%d", &x, &y, &z, &w) == 4) {
+  double x, y, z, w;
+  if (sscanf(s, "%lf,%lf,%lf,%lf", &x, &y, &z, &w) == 4) {
     v = boost::any(Quaternion(x, y, z, w));
   } else {
     throw validation_error(validation_error::invalid_option_value, "Camera rotation must be given as \"X,Y,Z,W\"");
